requirements.txt
flask
python-telegram-bot
movies-scraper